Move the bulk of creating of the settings page to a new component
Let AdminSettings make use this new component as well and gain the safely creation too
