+++ src/js/pages/adminSettings.js
@@ -1,20 +1,14 @@
 import {
   div,
-  label,
-  input,
   button,
-  h4,
-  span,
-  ul,
-  li,
 } from "@cycle/dom"
 import xs from "xstream"
 import isolate from "@cycle/isolate"
-import { mergeWith, merge, props, keys, cond } from "ramda"
 import * as R from "ramda"
-import { pick, mix } from "cycle-onionify"
 import debounce from "xstream/extra/debounce"
 import dropRepeats from "xstream/extra/dropRepeats"
+
+import { SettingsEditor } from "../components/SettingsEditor"
 
 export function IsolatedAdminSettings(sources) {
   return isolate(AdminSettings, "settings")(sources)
@@ -224,164 +218,7 @@ export function AdminSettings(sources) {
     },
   ])
 
-  // Depending on the type of config settings, render the appropriate vdom representation
-  function renderField(config, _state) {
-    if (config.type == "checkbox") {
-      return [
-        label(".active", [
-          input({ props: merge(config.props, { checked: _state }) }),
-          span(".lever"),
-        ]),
-      ]
-    }
-    if (config.type == "text") {
-      return [input({ props: merge(config.props, { value: _state }) })]
-    }
-    if (config.type == "select") {
-      const options = config.options
-      const selectedOption = (option) =>
-        _state == option
-          ? ".grey.lighten-3.black-text"
-          : ".grey.lighten-3.grey-text.text-lighten-1"
-      const optionButtons = options.map((o) =>
-        div(
-          ".col.selection" + selectedOption(o) + "." + o,
-          { style: { "border-style": "solid", margin: "2px" } },
-          [
-            label(selectedOption(o), [
-              input("", { props: merge(config.props, { value: o }) }, ""),
-              o,
-            ]),
-          ]
-        )
-      )
-      return optionButtons
-    }
-  }
-
-  const makeSetting = (config) => (sources) => {
-    const state$ = sources.onion.state$
-
-    const update$ = R.cond([
-      [
-        R.equals("checkbox"),
-        (_) =>
-          sources.DOM.select("input")
-            .events("click")
-            .map((event) => event),
-      ],
-      [
-        R.equals("text"),
-        (_) => sources.DOM.events("input").map((event) => event.target.value),
-      ],
-      [
-        R.equals("select"),
-        (_) =>
-          sources.DOM.select("input")
-            .events("click")
-            .map((event) => event.target.value),
-      ],
-    ])(config.type)
-
-    const vdom$ = state$.map((state) =>
-      li(
-        ".collection-item .row",
-        div(".valign-wrapper", [
-          span(".col .l6 .s12 .truncate", [
-            span(".flow-text", [config.title]),
-            span(["  "]),
-            span(".grey-text .text-lighten-1 .right-align", [
-              "(",
-              state.toString(),
-              ")",
-            ]),
-          ]),
-
-          div(".col .s6 " + config.class, renderField(config, state)),
-        ])
-      )
-    )
-
-    const updateReducer$ = cond([
-      [
-        R.equals("checkbox"),
-        (_) => update$.map((_) => (prevState) => !prevState),
-      ],
-      [R.equals("text"), (_) => update$.map((update) => (_) => update)],
-      [
-        R.equals("select"),
-        (_) => update$.map((update) => (_) => update),
-      ],
-    ])(config.type)
-
-    return {
-      DOM: vdom$,
-      onion: updateReducer$,
-    }
-  }
-
-  const makeSettingsGroup = (settingsGroupObj) => (sources) => {
-    const group$ = sources.onion.state$
-    const settingsArray = settingsGroupObj.settings
-    const title = settingsGroupObj.title
-
-    const components$ = xs
-      .of(settingsArray)
-      .map((settings) =>
-        settings.map((setting) =>
-          isolate(makeSetting(setting), setting.field)(sources)
-        )
-      )
-      .remember()
-
-    const vdom$ = components$
-      .compose(pick("DOM"))
-      .compose(mix(xs.combine))
-      .map((vdoms) =>
-        ul(
-          ".collection .with-header",
-          [li(".collection-header .grey .lighten-2", [h4(title)])].concat(vdoms)
-        )
-      )
-      .remember()
-
-    const reducer$ = components$.compose(pick("onion")).compose(mix(xs.merge))
-
-    return {
-      onion: reducer$,
-      DOM: vdom$,
-    }
-  }
-
-  const makeSettings = (settingsConf$, sources) => {
-    const settings$ = sources.onion.state$
-
-    const groups$ = settingsConf$
-      .map((groups) =>
-        groups.map((group) =>
-          isolate(makeSettingsGroup(group), group.group)(sources)
-        )
-      )
-      .remember()
-
-    const vdom$ = groups$
-      .compose(pick("DOM"))
-      .compose(mix(xs.combine))
-      .map((vdoms) => div(".col .l8 .offset-l2 .s12", vdoms))
-      .remember()
-
-    const reducer$ = groups$
-      .compose(pick("onion"))
-      .compose(mix(xs.merge))
-      .remember()
-
-    return {
-      onion: reducer$,
-      DOM: vdom$,
-    }
-  }
-
-  const AdminSettings = makeSettings(settingsConfig$, sources)
+  const AdminSettings = SettingsEditor({...sources, settings$: settingsConfig$})
 
   const vdom$ = xs
     .combine(settings$, AdminSettings.DOM)
